fix the accordion where now it can have transition on easing out
also move the accordion to another file to in UI folder under Accordion so the StockItem.jsx look more clean.
