+++ src/components/UI/Accordion/Accordion.module.scss
@@ -1,0 +1,42 @@
+@use "../../../styles/variables.scss";
+@use "../../../styles/mixins.scss";
+
+// vector svg icon to rotate on click
+.rotated {
+  transform: rotate(180deg);
+}
+
+// accordion styles
+.stockInfoContent {
+  flex-grow: 1;
+  flex-shrink: 1;
+  justify-content: center;
+  // /* to prevent content from overflowing during animation */
+  // transition: height 0.50s ease-in-out; //currently its only ease-in but not out
+  /* add a transition to animate height changes */
+  margin-top: 0.5rem;
+  // set max height of closing accordion to be 0
+  max-height: 0;
+  overflow: hidden;
+  transition: max-height .45s ease-in-out;
+}
+
+.stockInfoContent.open {
+  max-height: 100rem;
+  /* set max height of accordion when it open to 100rem */
+}
+
+.stockInfoText {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  background-color: #ffffff;
+  padding: 2rem;
+  border-bottom-left-radius: 25px;
+  border-bottom-right-radius: 25px;
+  border-radius: 25px;
+  font-size: 14px;
+  line-height: 1.5rem;
+  text-align: justify;
+}
